# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Academia",
        "description": "Sistema de gerenciamento de Academia",
        "version": "1.0.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "diegomunizdev@outlook.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "localhost HTTP"
        }
    ],
    "basePath": "http://localhost:3001/gym",
...
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/auth/user/{id}/changepassword": {
            "patch": {
                "tags": [
                    "Authentication"
                ],
                "summary": "",
                "description": "",
                "operationId": "ChangePassword",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Created user object",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "petstore_auth": {
            "type": "oauth2",
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "write:pets": "modify pets in your account",
                "read:pets": "read your pets"
            }
        },
        "api_key": {
            "type": "apiKey",
            "name": "api_key",
            "in": "header"
        }
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "required": [
                "id",
                "created_at",
                "username",
                "email",
                "password",
                "cpf",
                "type"
            ],
            "properties": {
                "id": {
                    "readOnly": true,
                    "type": "string",
                    "format": "uuid",
                    "example": "507f191e810c19729de860ea"
                },
                "created_at": {
                    "readOnly": true,
                    "type": "string",
                    "format": "date-time",
                    "description": "Date automatically generated in UTC format according to ISO 8601 at the time the resource is saved.",
                    "example": "2020-08-05T10:00:30Z"
                },
                "username": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "cpf": {
                    "type": "string"
                },
                "age": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "client",
                        "personal_trainer"
                    ]
                },
                "phone": {
                    "type": "string"
                },
                "genre": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ]
                },
                "contract_start": {
                    "type": "string"
                },
                "contract_end": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "User"
            }
        },
...
    },
...
}